fix: 유저 관계 설정 1차
