+++ backend/src/main/java/com/techeer/backend/api/feedback/dto/response/AllFeedbackResponse.java
@@ -1,4 +1,4 @@
-package com.techeer.backend.api.feedback.dto;
+package com.techeer.backend.api.feedback.dto.response;
 
 import java.util.List;
 import lombok.AllArgsConstructor;
